[♻️refactor]: postOrderPurchase 분리
